modularize get started sample (#86)
