+++ MultipazGettingStartedSample/composeApp/src/androidMain/kotlin/org/multipaz/getstarted/CredmanActivity.kt
@@ -2,13 +2,14 @@ package org.multipaz.getstarted
 
 import multipazgettingstartedsample.composeapp.generated.resources.Res
 import org.multipaz.compose.digitalcredentials.CredentialManagerPresentmentActivity
+import org.multipaz.getstarted.core.AppContainer
 
 class CredmanActivity : CredentialManagerPresentmentActivity() {
     override suspend fun getSettings(): Settings {
-        val app = App.getInstance()
-        app.init()
+        val container = AppContainer.getInstance()
+        container.init()
         return Settings(
-            source = app.presentmentSource,
+            source = container.presentmentSource,
             privilegedAllowList = Res.readBytes("files/privilegedUserAgents.json").decodeToString()
         )
     }
