--- MultipazGettingStartedSample/composeApp/src/androidMain/kotlin/org/multipaz/getstarted/Platform.kt
@@ -1,25 +1,0 @@
-package org.multipaz.getstarted
-
-import android.content.pm.PackageManager
-import io.ktor.client.engine.HttpClientEngineFactory
-import io.ktor.client.engine.android.Android
-import org.multipaz.context.applicationContext
-import org.multipaz.digitalcredentials.getAppOrigin
-
-@Suppress("DEPRECATION")
-actual suspend fun getAppToAppOrigin(): String {
-    val packageInfo = applicationContext.packageManager
-        .getPackageInfo(applicationContext.packageName, PackageManager.GET_SIGNATURES)
-
-    val signatures = packageInfo.signatures
-    if (signatures.isNullOrEmpty()) {
-        throw IllegalStateException("No signatures found for package ${applicationContext.packageName}")
-    }
-    return getAppOrigin(signatures[0].toByteArray())
-}
-
-actual fun isAndroid(): Boolean = true
-
-actual val httpClientEngineFactory: HttpClientEngineFactory<*> by lazy {
-    Android
-}
