+++ MultipazGettingStartedSample/composeApp/src/commonMain/kotlin/org/multipaz/getstarted/HomeScreen.kt
@@ -1,13 +1,11 @@
 package org.multipaz.getstarted
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.gestures.Orientation
 import androidx.compose.foundation.gestures.scrollable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material.icons.Icons
@@ -18,7 +16,6 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -26,7 +23,6 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
@@ -35,64 +31,45 @@ import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.io.bytestring.ByteString
 import kotlinx.serialization.json.JsonObject
 import org.multipaz.cbor.DataItem
-import org.multipaz.compose.camera.Camera
-import org.multipaz.compose.camera.CameraCaptureResolution
-import org.multipaz.compose.camera.CameraFrame
-import org.multipaz.compose.camera.CameraSelection
-import org.multipaz.compose.decodeImage
-import org.multipaz.compose.permissions.rememberBluetoothEnabledState
-import org.multipaz.compose.permissions.rememberBluetoothPermissionState
 import org.multipaz.compose.permissions.rememberCameraPermissionState
-import org.multipaz.compose.presentment.MdocProximityQrPresentment
-import org.multipaz.compose.presentment.MdocProximityQrSettings
-import org.multipaz.compose.qrcode.generateQrCode
 import org.multipaz.crypto.EcPrivateKey
 import org.multipaz.document.Document
-import org.multipaz.facedetection.detectFaces
+import org.multipaz.getstarted.biometrics.FaceExtractor
+import org.multipaz.getstarted.biometrics.FaceMatchingSection
+import org.multipaz.getstarted.biometrics.SelfieCheckSection
+import org.multipaz.getstarted.core.AppContainer
+import org.multipaz.getstarted.core.CredentialDomains
+import org.multipaz.getstarted.core.isAndroid
+import org.multipaz.getstarted.presentment.PresentmentHomeSection
+import org.multipaz.getstarted.verification.ShowResponseMetadata
+import org.multipaz.getstarted.verification.W3CDCCredentialsRequestButton
+import org.multipaz.getstarted.verification.buildShowResponseDestination
 import org.multipaz.facematch.FaceEmbedding
-import org.multipaz.facematch.getFaceEmbeddings
-import org.multipaz.getstarted.App.Companion.SAMPLE_DOCUMENT_DISPLAY_NAME
-import org.multipaz.getstarted.w3cdc.ShowResponseMetadata
-import org.multipaz.getstarted.w3cdc.W3CDCCredentialsRequestButton
-import org.multipaz.getstarted.w3cdc.buildShowResponseDestination
-import org.multipaz.mdoc.connectionmethod.MdocConnectionMethod
-import org.multipaz.mdoc.connectionmethod.MdocConnectionMethodBle
-import org.multipaz.mdoc.transport.MdocTransportOptions
-import org.multipaz.selfiecheck.SelfieCheck
-import org.multipaz.selfiecheck.SelfieCheckViewModel
-import org.multipaz.util.UUID
-import kotlin.math.roundToInt
-import kotlin.time.Duration.Companion.seconds
 
 @Composable
 fun HomeScreen(
-    app: App,
+    container: AppContainer,
     navController: NavController,
     documents: List<Document>,
     identityIssuer: String = "Multipaz Getting Started Sample",
     onDeleteDocument: (Document) -> Unit
 ) {
-    val selfieCheckViewModel: SelfieCheckViewModel =
-        remember { SelfieCheckViewModel(identityIssuer) }
-
-    val blePermissionState = rememberBluetoothPermissionState()
-    val bleEnabledState = rememberBluetoothEnabledState()
-    val coroutineScope = rememberCoroutineScope { App.promptModel }
-
+    val coroutineScope = rememberCoroutineScope { AppContainer.promptModel }
     val uriHandler = LocalUriHandler.current
-
     val cameraPermissionState = rememberCameraPermissionState()
 
-    var showCamera by remember { mutableStateOf(false) }
+    val faceExtractor = remember { FaceExtractor() }
+    var faceExtractorReady by remember { mutableStateOf(false) }
+    LaunchedEffect(Unit) {
+        faceExtractor.init()
+        faceExtractorReady = true
+    }
+
     val faceCaptured = remember { mutableStateOf<FaceEmbedding?>(null) }
-    var showFaceMatching by remember { mutableStateOf(false) }
-    var similarity by remember { mutableStateOf(0f) }
     val scrollState = rememberScrollState()
 
     Column(
@@ -106,150 +83,69 @@ fun HomeScreen(
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        if (!blePermissionState.isGranted) {
-            Button(
-                onClick = {
-                    coroutineScope.launch {
-                        blePermissionState.launchPermissionRequest()
+        PresentmentHomeSection(
+            presentmentSource = container.presentmentSource,
+            promptModel = AppContainer.promptModel,
+            modifier = Modifier.weight(1f)
+        )
+
+        Button(
+            modifier = Modifier.padding(16.dp),
+            onClick = {
+                uriHandler.openUri("https://issuer.multipaz.org")
+            }) {
+            Text(
+                buildAnnotatedString {
+                    withStyle(style = SpanStyle(fontSize = 14.sp)) {
+                        append("Issue an mDoc from the server")
+                    }
+                    withStyle(style = SpanStyle(fontSize = 12.sp)) {
+                        append("\nhttps://issuer.multipaz.org")
+                    }
+                },
+                textAlign = TextAlign.Center
+            )
+        }
+
+        if (documents.isNotEmpty()) {
+            Text(
+                modifier = Modifier.padding(4.dp),
+                text = "${documents.size} Documents present:"
+            )
+            documents.forEachIndexed { index, document ->
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    Text(
+                        text = document.displayName ?: document.identifier,
+                        modifier = Modifier.padding(4.dp)
+                    )
+                    if (document.displayName != CredentialDomains.SAMPLE_DOCUMENT_DISPLAY_NAME) {
+                        IconButton(
+                            content = @Composable {
+                                Icon(
+                                    imageVector = Icons.Default.Delete,
+                                    contentDescription = null
+                                )
+                            },
+                            onClick = {
+                                coroutineScope.launch {
+                                    container.documentStore.deleteDocument(document.identifier)
+                                    onDeleteDocument(document)
+                                }
+                            }
+                        )
                     }
                 }
-            ) {
-                Text("Request BLE permissions")
-            }
-        } else if (!bleEnabledState.isEnabled) {
-            Button(
-                onClick = { coroutineScope.launch { bleEnabledState.enable() } }) {
-                Text("Enable Bluetooth")
             }
         } else {
-
-            MdocProximityQrPresentment(
-                modifier = Modifier.weight(1f),
-                source = app.presentmentSource,
-                promptModel = App.promptModel,
-                prepareSettings = { generateQrCode ->
-                    val connectionMethods = mutableListOf<MdocConnectionMethod>()
-                    val bleUuid = UUID.randomUUID()
-                    connectionMethods.add(
-                        MdocConnectionMethodBle(
-                            supportsPeripheralServerMode = true,
-                            supportsCentralClientMode = false,
-                            peripheralServerModeUuid = bleUuid,
-                            centralClientModeUuid = null,
-                        )
-                    )
-
-                    Column(
-                        modifier = Modifier.fillMaxSize().padding(16.dp),
-                        verticalArrangement = Arrangement.Center,
-                        horizontalAlignment = Alignment.CenterHorizontally,
-                    ) {
-                        Button(
-                            onClick = {
-                                generateQrCode(
-                                    MdocProximityQrSettings(
-                                        availableConnectionMethods = connectionMethods,
-                                        createTransportOptions = MdocTransportOptions(
-                                            bleUseL2CAPInEngagement = true
-                                        )
-                                    )
-                                )
-                            }
-                        ) @Composable {
-                            Text("Present mDoc via QR code")
-                        }
-                    }
-                },
-                showTransacting = { reset ->
-                    Text("Transacting")
-                    Button(onClick = { reset() }) {
-                        Text("Cancel")
-                    }
-                },
-                showQrCode = { uri, reset ->
-                    ShowQrCode(
-                        uri,
-                        onCancel = {
-                            reset()
-                        }
-                    )
-                },
-                showCompleted = { error, reset ->
-                    if (error is CancellationException) {
-                        reset()
-                    } else {
-                        if (error != null) {
-                            Text("Something went wrong: $error")
-                        } else {
-                            Text("The data was shared")
-                        }
-                        LaunchedEffect(Unit) {
-                            delay(1.5.seconds)
-                            reset()
-                        }
-                    }
-                },
-            )
-
-            Button(
-                modifier = Modifier.padding(16.dp),
-                onClick = {
-                    uriHandler.openUri("https://issuer.multipaz.org")
-                }) {
-                Text(
-                    buildAnnotatedString {
-                        withStyle(style = SpanStyle(fontSize = 14.sp)) {
-                            append("Issue an mDoc from the server")
-                        }
-                        withStyle(style = SpanStyle(fontSize = 12.sp)) {
-                            append("\nhttps://issuer.multipaz.org")
-                        }
-                    },
-                    textAlign = TextAlign.Center
-                )
-            }
-
-
-            if (documents.isNotEmpty()) {
-                Text(
-                    modifier = Modifier.padding(4.dp),
-                    text = "${documents.size} Documents present:"
-                )
-                documents.forEachIndexed { index, document ->
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text(
-                            text = document.displayName ?: document.identifier,
-                            modifier = Modifier.padding(4.dp)
-                        )
-                        if (document.displayName != SAMPLE_DOCUMENT_DISPLAY_NAME) {
-                            IconButton(
-                                content = @Composable {
-                                    Icon(
-                                        imageVector = Icons.Default.Delete,
-                                        contentDescription = null
-                                    )
-                                },
-                                onClick = {
-                                    coroutineScope.launch {
-                                        app.documentStore.deleteDocument(document.identifier)
-                                        onDeleteDocument(document)
-                                    }
-                                }
-                            )
-                        }
-                    }
-                }
-            } else {
-                Text(text = "No documents found.")
-            }
+            Text(text = "No documents found.")
         }
 
         // W3C Digital Credentials API is only available on Android
         if (isAndroid() && documents.isNotEmpty()) {
             W3CDCCredentialsRequestButton(
-                promptModel = App.promptModel,
-                storageTable = app.storageTable,
-                readerTrustManager = app.readerTrustManager,
+                promptModel = AppContainer.promptModel,
+                storageTable = container.storageTable,
+                readerTrustManager = container.readerTrustManager,
                 showResponse = { vpToken: JsonObject?,
                                  deviceResponse: DataItem?,
                                  sessionTranscript: DataItem,
@@ -270,199 +166,37 @@ fun HomeScreen(
             )
         }
 
-        when {
-            !cameraPermissionState.isGranted -> {
-                Button(
-                    onClick = {
-                        coroutineScope.launch {
-                            cameraPermissionState.launchPermissionRequest()
-                        }
-                    }
-                ) {
-                    Text("Grant Camera Permission for Selfie Check")
-                }
-            }
-
-            faceCaptured.value == null -> {
-                SelfieCheckFlow(
-                    showCamera = showCamera,
-                    onShowCameraChange = { showCamera = it },
-                    selfieCheckViewModel = selfieCheckViewModel,
-                    identityIssuer = identityIssuer,
-                    onFaceCaptured = { embedding ->
-                        faceCaptured.value = embedding
-                    },
-                    app = app
-                )
-            }
-
-            else -> {
-                FaceMatchingFlow(
-                    showFaceMatching = showFaceMatching,
-                    onShowFaceMatchingChange = { showFaceMatching = it },
-                    similarity = similarity,
-                    onSimilarityChange = { similarity = it },
-                    faceCaptured = faceCaptured,
-                    app = app
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun SelfieCheckFlow(
-    showCamera: Boolean,
-    onShowCameraChange: (Boolean) -> Unit,
-    selfieCheckViewModel: SelfieCheckViewModel,
-    identityIssuer: String,
-    onFaceCaptured: (FaceEmbedding?) -> Unit,
-    app: App
-) {
-    if (!showCamera) {
-        Button(onClick = { onShowCameraChange(true) }) {
-            Text("Selfie Check")
-        }
-    } else {
-        SelfieCheck(
-            modifier = Modifier.fillMaxWidth(),
-            onVerificationComplete = {
-                onShowCameraChange(false)
-                if (selfieCheckViewModel.capturedFaceImage != null) {
-                    val embedding = getFaceEmbeddings(
-                        image = decodeImage(selfieCheckViewModel.capturedFaceImage!!.toByteArray()),
-                        model = app.faceMatchLiteRtModel
-                    )
-                    onFaceCaptured(embedding)
-                }
-                selfieCheckViewModel.resetForNewCheck()
-            },
-            viewModel = selfieCheckViewModel,
-            identityIssuer = identityIssuer
-        )
-
-        Button(
-            onClick = {
-                onShowCameraChange(false)
-                selfieCheckViewModel.resetForNewCheck()
-            }
-        ) {
-            Text("Close")
-        }
-    }
-}
-
-@Composable
-private fun FaceMatchingFlow(
-    showFaceMatching: Boolean,
-    onShowFaceMatchingChange: (Boolean) -> Unit,
-    similarity: Float,
-    onSimilarityChange: (Float) -> Unit,
-    faceCaptured: MutableState<FaceEmbedding?>,
-    app: App
-) {
-    if (!showFaceMatching) {
-        Button(onClick = { onShowFaceMatchingChange(true) }) {
-            Text("Face Matching")
-        }
-    } else {
-        Text("Similarity: ${(similarity * 100).roundToInt()}%")
-
-        Camera(
-            modifier = Modifier
-                .fillMaxSize(0.5f)
-                .padding(64.dp),
-            cameraSelection = CameraSelection.DEFAULT_FRONT_CAMERA,
-            captureResolution = CameraCaptureResolution.MEDIUM,
-            showCameraPreview = true,
-        ) { incomingVideoFrame: CameraFrame ->
-            val faces = detectFaces(incomingVideoFrame)
-
+        if (faceExtractorReady) {
             when {
-                faces.isNullOrEmpty() -> {
-                    onSimilarityChange(0f)
-                }
-
-                faceCaptured.value != null -> {
-                    val faceImage = app.extractFaceBitmap(
-                        incomingVideoFrame,
-                        faces[0], // assuming only one face exists for simplicity
-                        app.faceMatchLiteRtModel.imageSquareSize
-                    )
-
-                    val faceEmbedding = getFaceEmbeddings(faceImage, app.faceMatchLiteRtModel)
-
-                    if (faceEmbedding != null) {
-                        val newSimilarity = faceCaptured.value!!.calculateSimilarity(faceEmbedding)
-                        onSimilarityChange(newSimilarity)
+                !cameraPermissionState.isGranted -> {
+                    Button(
+                        onClick = {
+                            coroutineScope.launch {
+                                cameraPermissionState.launchPermissionRequest()
+                            }
+                        }
+                    ) {
+                        Text("Grant Camera Permission for Selfie Check")
                     }
                 }
-            }
-        }
 
-        Button(
-            onClick = {
-                onShowFaceMatchingChange(false)
-                faceCaptured.value = null
-            }
-        ) {
-            Text("Close")
-        }
-    }
-}
+                faceCaptured.value == null -> {
+                    SelfieCheckSection(
+                        faceExtractor = faceExtractor,
+                        identityIssuer = identityIssuer,
+                        onFaceCaptured = { embedding ->
+                            faceCaptured.value = embedding
+                        }
+                    )
+                }
 
-@Composable
-fun ShowQrButton(onQrButtonClicked: (settings: MdocProximityQrSettings) -> Unit) {
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Button(onClick = {
-            val connectionMethods = listOf(
-                MdocConnectionMethodBle(
-                    supportsPeripheralServerMode = false,
-                    supportsCentralClientMode = true,
-                    peripheralServerModeUuid = null,
-                    centralClientModeUuid = UUID.randomUUID(),
-                )
-            )
-            onQrButtonClicked(
-                MdocProximityQrSettings(
-                    availableConnectionMethods = connectionMethods,
-                    createTransportOptions = MdocTransportOptions(bleUseL2CAP = true)
-                )
-            )
-        }) {
-            Text("Present mDL via QR Code")
-        }
-    }
-}
-
-@Composable
-fun ShowQrCode(
-    uri: String,
-    onCancel: () -> Unit
-) {
-    Column(
-        modifier = Modifier.fillMaxSize().padding(16.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        val qrCodeBitmap = remember { generateQrCode(uri) }
-        Text(text = "Present QR code to mdoc reader")
-        Image(
-            modifier = Modifier.fillMaxWidth(),
-            bitmap = qrCodeBitmap,
-            contentDescription = null,
-            contentScale = ContentScale.FillWidth
-        )
-        Button(
-            onClick = {
-                onCancel()
+                else -> {
+                    FaceMatchingSection(
+                        faceExtractor = faceExtractor,
+                        faceCaptured = faceCaptured
+                    )
+                }
             }
-        ) {
-            Text("Cancel")
         }
     }
 }
