+++ MultipazGettingStartedSample/composeApp/src/commonMain/kotlin/org/multipaz/getstarted/Navigation.kt
@@ -9,14 +9,4 @@ sealed interface Destination {
     data object HomeDestination : Destination
     @Serializable
     data object ProvisioningDestination : Destination
-
-    @Serializable
-    data class ShowResponseDestination(
-        val vpResponse: String?,
-        val deviceResponse: String?,
-        val sessionTranscript: String?,
-        val nonce: String?,
-        val eReaderKey: String?,
-        val metadata: String?
-    ) : Destination
 }
