--- MultipazGettingStartedSample/composeApp/src/commonMain/kotlin/org/multipaz/getstarted/Platform.kt
@@ -1,9 +1,0 @@
-package org.multipaz.getstarted
-
-import io.ktor.client.engine.HttpClientEngineFactory
-
-expect suspend fun getAppToAppOrigin(): String
-
-expect fun isAndroid(): Boolean
-
-expect val httpClientEngineFactory: HttpClientEngineFactory<*>
