--- MultipazGettingStartedSample/composeApp/src/commonMain/kotlin/org/multipaz/getstarted/w3cdc/DocumentValue.kt
@@ -1,14 +1,0 @@
-package org.multipaz.getstarted.w3cdc
-
-import androidx.compose.ui.graphics.ImageBitmap
-
-sealed class DocumentValue {
-     data class ValueText(
-        val title: String,
-        val value: String
-    ): DocumentValue()
-
-    data class ValueImage(
-        val image: ImageBitmap
-    ): DocumentValue()
-}
