+++ MultipazGettingStartedSample/core/build.gradle.kts
@@ -1,0 +1,63 @@
+import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+
+plugins {
+    alias(libs.plugins.kotlinMultiplatform)
+    alias(libs.plugins.androidLibrary)
+    alias(libs.plugins.composeMultiplatform)
+    alias(libs.plugins.composeCompiler)
+    alias(libs.plugins.kotlinSerialization)
+}
+
+kotlin {
+    jvmToolchain(17)
+
+    androidTarget {
+        @OptIn(ExperimentalKotlinGradlePluginApi::class)
+        compilerOptions {
+            jvmTarget.set(JvmTarget.JVM_17)
+        }
+    }
+
+    iosX64()
+    iosArm64()
+    iosSimulatorArm64()
+    applyDefaultHierarchyTemplate()
+
+    sourceSets {
+        androidMain.dependencies {
+            implementation(libs.ktor.client.android)
+        }
+        commonMain.dependencies {
+            implementation(compose.runtime)
+            implementation(compose.components.resources)
+            implementation(libs.multipaz)
+            implementation(libs.multipaz.compose)
+            implementation(libs.multipaz.doctypes)
+            implementation(libs.multipaz.dcapi)
+            implementation(libs.multipaz.vision)
+            implementation(libs.ktor.client.core)
+            implementation(libs.ktor.client.cio)
+            implementation(libs.kotlinx.serialization.json)
+            implementation(libs.androidx.lifecycle.viewmodel)
+        }
+        iosMain.dependencies {
+            implementation(libs.ktor.client.darwin)
+            implementation(libs.androidx.sqlite)
+            implementation(libs.androidx.sqlite.framework)
+        }
+    }
+}
+
+android {
+    namespace = "org.multipaz.getstarted.core"
+    compileSdk = libs.versions.android.compileSdk.get().toInt()
+
+    defaultConfig {
+        minSdk = libs.versions.android.minSdk.get().toInt()
+    }
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+}
