+++ MultipazGettingStartedSample/core/src/commonMain/composeResources/files/reader_root_cert_multipaz_web_verifier.pem
