+++ MultipazGettingStartedSample/core/src/commonMain/kotlin/org/multipaz/getstarted/core/Platform.kt
@@ -1,0 +1,9 @@
+package org.multipaz.getstarted.core
+
+import io.ktor.client.engine.HttpClientEngineFactory
+
+expect suspend fun getAppToAppOrigin(): String
+
+expect fun isAndroid(): Boolean
+
+expect val httpClientEngineFactory: HttpClientEngineFactory<*>
