+++ MultipazGettingStartedSample/feature/presentment/src/commonMain/kotlin/org/multipaz/getstarted/presentment/QrCodeDisplay.kt
@@ -1,0 +1,77 @@
+package org.multipaz.getstarted.presentment
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.unit.dp
+import org.multipaz.compose.presentment.MdocProximityQrSettings
+import org.multipaz.compose.qrcode.generateQrCode
+import org.multipaz.mdoc.connectionmethod.MdocConnectionMethodBle
+import org.multipaz.mdoc.transport.MdocTransportOptions
+import org.multipaz.util.UUID
+
+@Composable
+fun ShowQrButton(onQrButtonClicked: (settings: MdocProximityQrSettings) -> Unit) {
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Button(onClick = {
+            val connectionMethods = listOf(
+                MdocConnectionMethodBle(
+                    supportsPeripheralServerMode = false,
+                    supportsCentralClientMode = true,
+                    peripheralServerModeUuid = null,
+                    centralClientModeUuid = UUID.randomUUID(),
+                )
+            )
+            onQrButtonClicked(
+                MdocProximityQrSettings(
+                    availableConnectionMethods = connectionMethods,
+                    createTransportOptions = MdocTransportOptions(bleUseL2CAP = true)
+                )
+            )
+        }) {
+            Text("Present mDL via QR Code")
+        }
+    }
+}
+
+@Composable
+fun ShowQrCode(
+    uri: String,
+    onCancel: () -> Unit
+) {
+    Column(
+        modifier = Modifier.fillMaxSize().padding(16.dp),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        val qrCodeBitmap = remember { generateQrCode(uri) }
+        Text(text = "Present QR code to mdoc reader")
+        Image(
+            modifier = Modifier.fillMaxWidth(),
+            bitmap = qrCodeBitmap,
+            contentDescription = null,
+            contentScale = ContentScale.FillWidth
+        )
+        Button(
+            onClick = {
+                onCancel()
+            }
+        ) {
+            Text("Cancel")
+        }
+    }
+}
