+++ MultipazGettingStartedSample/feature/verification/src/commonMain/composeResources/files/reader_root_key_private.pem
