+++ MultipazGettingStartedSample/feature/verification/src/commonMain/composeResources/files/reader_root_key_public.pem
