+++ MultipazGettingStartedSample/feature/verification/src/commonMain/kotlin/org/multipaz/getstarted/verification/DocumentValue.kt
@@ -1,0 +1,14 @@
+package org.multipaz.getstarted.verification
+
+import androidx.compose.ui.graphics.ImageBitmap
+
+sealed class DocumentValue {
+     data class ValueText(
+        val title: String,
+        val value: String
+    ): DocumentValue()
+
+    data class ValueImage(
+        val image: ImageBitmap
+    ): DocumentValue()
+}
