+++ MultipazGettingStartedSample/feature/verification/src/commonMain/kotlin/org/multipaz/getstarted/verification/ShowResponseDestination.kt
@@ -1,0 +1,13 @@
+package org.multipaz.getstarted.verification
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ShowResponseDestination(
+    val vpResponse: String?,
+    val deviceResponse: String?,
+    val sessionTranscript: String?,
+    val nonce: String?,
+    val eReaderKey: String?,
+    val metadata: String?
+)
